+++ noodles-tabix/src/io/writer/index.rs
@@ -15,6 +15,8 @@ where
 {
     use noodles_csi::io::writer::index::write_header;
 
+    validate(index)?;
+
     write_magic_number(writer)?;
 
     let reference_sequence_count = i32::try_from(index.reference_sequences().len())
@@ -33,6 +35,25 @@ where
     }
 
     Ok(())
+}
+
+fn validate(index: &Index) -> io::Result<()> {
+    const MIN_SHIFT: u8 = 14;
+    const DEPTH: u8 = 5;
+
+    let min_shift = index.min_shift();
+    let depth = index.depth();
+
+    if min_shift == MIN_SHIFT && depth == DEPTH {
+        Ok(())
+    } else {
+        Err(io::Error::new(
+            io::ErrorKind::InvalidInput,
+            format!(
+                "invalid (min shift, depth): expected ({MIN_SHIFT}, {DEPTH}), got ({min_shift}, {depth})"
+            ),
+        ))
+    }
 }
 
 #[cfg(test)]
@@ -95,5 +116,17 @@ mod tests {
         assert_eq!(buf, expected);
 
         Ok(())
+    }
+
+    #[test]
+    fn test_validate() {
+        let index = Index::default();
+        assert!(validate(&index).is_ok());
+
+        let index = Index::builder().set_min_shift(13).set_depth(8).build();
+        assert!(matches!(
+            validate(&index),
+            Err(e) if e.kind() == io::ErrorKind::InvalidInput
+        ));
     }
 }