+++ noodles-tabix/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## Unreleased
+
+### Changed
+
+  * tabix/io/writer/index: Validate min shift and depth.
+
+    These are now required to be the values assumed by tabix (14 and 5,
+    respectively).
+
 ## 0.67.0 - 2026-08-21
 
 ### Changed